+++ resources/views/component/navbar.blade.php
@@ -1,3 +1,3 @@
-@include('component.navbar.mini')
 @include('component.navbar.main')
 @include('component.navbar.menu')
+@include('component.navbar.mini')