+++ resources/views/layout/customer/type/input.blade.php
@@ -9,24 +9,24 @@
 @section('content')
     @include('component.breadcrumb')
 
+    @include('form.header.start')
     {{--    Body--}}
     <div class="row mt-3">
         <div class="col-md-3">
-            <h5 class="text-black mb-0">Main Menu</h5>
-            <small class="text-muted">Data ini yang nanti akan digunakan saat memberikan nama cabang</small>
+            <h5 class="text-black mb-0">Information</h5>
+            <small class="text-muted">Data ini yang nanti akan digunakan saat memberikan informasi tipe konsumen</small>
             <br><small class="text-danger">*Wajib diisi</small>
         </div>
         <div id="main" class="col-md-9">
-            @include('form.header.start')
             <div class="card mb-3">
                 <div class="card-body">
                     <div class="mb-3">
                         <label class="form-label">Name<span class="text-danger">*</span></label>
-                        <input type="text" name="name" class="form-control w-50 @error('name') is-invalid @enderror" value="{{ old('name', $data->name) }}" placeholder="Nama Cabang">
+                        <input type="text" name="name" class="form-control w-50 @error('name') is-invalid @enderror" value="{{ old('name', $data->name) }}" placeholder="Nama Tipe Konsumen">
                         @error('name')
-                        <div class="invalid-feedback">
-                            {{ $message }}
-                        </div>
+                            <div class="invalid-feedback">
+                                {{ $message }}
+                            </div>
                         @enderror
                     </div>
                 </div>
@@ -36,7 +36,7 @@
                     </div>
                 </div>
             </div>
-            @include('form.header.end')
         </div>
     </div>
+    @include('form.header.end')
 @endsection