+++ resources/views/layout/customer/type/table.blade.php
@@ -1,14 +1,12 @@
 @include('component.table.header', ['perPage' => $table['perPage'], 'search' => $table['search']])
-<div class="table-container">
+<div class="table-container table-responsive">
     <table class="table table-striped mb-0 data-table align-middle">
         <thead>
         <tr>
-            @include('component.table.title', ['title' => '#', 'column' => 'branch_coordinators.id', 'sortable' => true, 'class' => 'text-center'])
-            @include('component.table.title', ['title' => 'Name', 'column' => 'branch_coordinators.name', 'sortable' => true])
-            @include('component.table.title', ['title' => 'Coordinator', 'column' => 'users.name', 'sortable' => true])
-            @include('component.table.title', ['title' => 'Privilege', 'column' => 'user_privileges.name', 'sortable' => false, 'class' => 'align-middle'])
+            @include('component.table.title', ['title' => '#', 'column' => 'id', 'sortable' => true, 'class' => 'text-center'])
+            @include('component.table.title', ['title' => 'Name', 'column' => 'name', 'sortable' => true])
             {{-- Jika can CRUD maka munculkan tombol--}}
-            @if(getUserLevel('branches') >= CAN_CRUD)
+            @if(getUserLevel($config['privilege']) >= CAN_CRUD)
                 @include('component.table.title', ['title' => 'Action', 'column' => 'action', 'sortable'=> false, 'class' => 'text-center align-middle'])
             @endif
         </tr>
@@ -17,8 +15,8 @@
         @if($data->total() == 0)
             <tr>
                 @php
-                    $colspan = 4;
-                    if (getUserLevel('branches') >= CAN_CRUD) {
+                    $colspan = 2;
+                    if (getUserLevel($config['privilege']) >= CAN_CRUD) {
                         $colspan += 1; // ada baguian untuk action button
                     }
                 @endphp
@@ -29,25 +27,11 @@
                 <tr>
                     <td class="ps-3 text-muted w-1-slot">{{ $table['column'] == 'id' && $table['sort'] == 'desc' ? $data->total() - ($data->firstItem() + $key) + 1 : $data->firstItem() + $key }}</td>
                     <td>{{ $row->name }}</td>
-                    <td>
-                        @if($row->user)
-                            <span>{{ $row->users->name }}</span>
-                        @else
-                            <span>-</span>
-                        @endif
-                    </td>
-                    <td>
-                        @if($row->user)
-                            <span class="{{ getBadge($row->users->privileges->color) }}">{{ $row->users->privileges->name }}</span>
-                        @else
-                            <span>-</span>
-                        @endif
-                    </td>
                     {{--                                    Jika can CRUD maka munculkan tombol--}}
                     @if(getUserLevel('branches') >= CAN_CRUD)
                         <td class="pe-3 w-2-slot">
                             <div class="d-flex">
-                                @include('form.button.crud', ['url' => 'branch/coordinator/', 'type' => $type, 'id' => $row->id])
+                                @include('form.button.crud', ['url' => $config['url'].'/', 'type' => $type, 'id' => $row->id])
                             </div>
                         </td>
                     @endif